✨ 添加蛋组

+++ avalon-modules/avalon-dataset/avalon-dataset-repository/src/main/java/io/github/lishangbu/avalon/dataset/repository/GenerationRepository.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Repository;
  * @since 2025/4/14
  */
 @Repository
-public interface GenerationRepository extends JpaRepository<Generation, Long> {
+public interface GenerationRepository extends JpaRepository<Generation, Integer> {
 
   Optional<Generation> findByName(String name);
 }
